# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tunnex API
  version: 0.1.0
  description: |
    Control-plane API for Tunnex.io. This spec is the single source of truth:
    the Go server types/handlers and the TypeScript client are generated from it
    (`make generate`), and drift fails the build (`make generate-check`).
servers:
  - url: /
    description: Same-origin (served behind the edge nginx)

# Endpoints require a session cookie OR a CLI bearer credential by default
# (bearer ≡ cookie, S5.1); public ones set `security: []`; cookie-only ops
# (credential minting) override with `security: [{cookieAuth: []}]` + argument.
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/agent/runtime/poll:
    get:
      operationId: pollAgentRuntime
      summary: Poll managed agent configuration
      description: >-
        Machine-only managed-runtime channel. Authenticates with the exact agent runtime bearer and returns
        a configuration only when the desired revision is newer than `applied_revision`; an unchanged
        revision returns 204 with no body. The optional wait is a bounded long poll. Explicit organization
        opt-out is a terminal uniform 401 so the runtime offboards; edition unavailability remains a
        non-terminal 403. No bearer, bootstrap token, private key, token hash, or raw error is ever returned.
      security:
        - runtimeBearer: []
      tags: [agents]
      parameters:
        - name: applied_revision
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        - name: client_version
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: wait_seconds
          in: query
          required: false
          description: Bounded long-poll wait requested by the runtime; the server may return sooner.
          schema: { type: integer, minimum: 0, maximum: 60, default: 0 }
      responses:
        "200":
          description: Strictly newer managed configuration.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedAgentConfig" }
        "204":
          description: Desired revision is unchanged; no configuration body is returned.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
        "401": { $ref: "#/components/responses/RuntimeUnauthorized" }
        "403": { $ref: "#/components/responses/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/agent/runtime/report:
    post:
      operationId: reportAgentRuntime
      summary: Report managed agent runtime state
      description: >-
        Machine-only managed-runtime channel. Reports bounded revision and health facts for the exact agent
        runtime bearer. The server rejects backwards or ahead-of-desired revisions and stores only the stable
        error enum, never raw runtime output. No credential, bootstrap token, private key, token hash, or raw
        error is accepted or returned.
      security:
        - runtimeBearer: []
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentRuntimeReport" }
      responses:
        "204":
          description: Report accepted.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
        "401": { $ref: "#/components/responses/RuntimeUnauthorized" }
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/organizations/{orgId}/agents/{deviceId}/runtime-status:
    parameters:
      - name: orgId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: deviceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getAgentRuntimeStatus
      summary: Get managed agent runtime status
      description: >-
        Organization/admin read-only projection for one agent. Uses human/session authorization, never the
        machine runtime bearer. The projection contains desired/applied facts, connectivity/last-seen state,
        and bounded stale/error facts only; it never returns credentials, private keys, token hashes, or raw
        runtime errors.
      tags: [agents]
      responses:
        "200":
          description: Runtime status projection.
          headers: { X-Request-Id: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentRuntimeStatus" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: tunnex_session
    # CLI credential (tnx_...). bearer ≡ cookie on every authenticated endpoint;
    # the ONLY exceptions are argued on the ops themselves (cliAuthorize +
    # cliDeviceApprove — credential minting must not be reachable from a bearer).
    bearerAuth:
      type: http
      scheme: bearer
    runtimeBearer:
      type: http
      scheme: bearer
      description: Machine-only hash-backed managed-agent runtime credential; never a user/session bearer.
  headers:
    RequestId:
      description: Correlation ID for this request; also appears in server logs.
      schema:
        type: string
      required: true
  schemas:
...
    ManagedAgentConfig:
      type: object
      description: >-
        Secret-free server-owned configuration for one managed agent. Runtime credentials, bootstrap tokens,
        token hashes, and private keys are never configuration fields.
      required: [revision, device_id, org_id, address, gateway_endpoint, gateway_public_key, allowed_ips, dns, persistent_keepalive]
      additionalProperties: false
      properties:
        revision: { type: integer, format: int64, minimum: 1 }
        device_id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        address: { type: string, description: The managed agent tunnel address/prefix. }
        gateway_endpoint: { type: string, description: The gateway endpoint in host:port form. }
        gateway_public_key: { type: string, description: The gateway WireGuard public key. }
        allowed_ips:
          type: array
          items: { type: string }
        dns:
          type: array
          items: { type: string }
        persistent_keepalive: { type: integer, minimum: 0, maximum: 65535 }
...
    AgentRuntimeReport:
      type: object
      description: Bounded, secret-free facts reported by one managed agent runtime.
      required: [applied_revision, attempted_revision, client_version, error_code]
      additionalProperties: false
      properties:
        applied_revision: { type: integer, format: int64, minimum: 0 }
        attempted_revision: { type: integer, format: int64, minimum: 0 }
        client_version: { type: string, minLength: 1, maxLength: 128 }
        error_code:
          type: string
          enum: ["", invalid_config, apply_failed]
          description: Empty means no error; otherwise a bounded stable code, never raw runtime output.
    AgentRuntimeStatus:
      type: object
      description: Secret-free organization/admin projection of one managed agent runtime.
      required: [device_id, desired_revision, applied_revision, last_attempted_revision, client_version, connectivity, health, stale]
      additionalProperties: false
      properties:
        device_id: { type: string, format: uuid }
        desired_revision: { type: integer, format: int64, minimum: 1 }
        applied_revision: { type: integer, format: int64, minimum: 0 }
        last_attempted_revision: { type: integer, format: int64, minimum: 0 }
        client_version: { type: string, maxLength: 128 }
        connectivity:
          type: string
          enum: [unknown, connected, disconnected]
          description: Server-derived connectivity fact; unknown means the control plane cannot establish it.
        health:
          type: string
          enum: [ready, last_good, inconclusive]
          description: >-
            ready means a fresh report applied the current desired revision without a bounded error;
            last_good means a prior applied revision is retained while freshness, revision, or apply state
            is not current; inconclusive means no applied configuration is established.
        stale:
          type: boolean
          description: Whether the last runtime report is absent or outside the server's three-minute freshness window.
        last_seen_at: { type: string, format: date-time, nullable: true }
        last_error_code:
          type: string
          nullable: true
          enum: [invalid_config, apply_failed]
          description: Last bounded stable runtime error, never raw runtime output.
        last_error_revision: { type: integer, format: int64, minimum: 1, nullable: true }
...
    Error:
      type: object
      description: |
        Standard error envelope for every non-2xx response. Chosen over bare
        RFC 7807 so the correlation `request_id` is a first-class field the SPA
        and CLI can surface directly; `code` is stable for programmatic handling.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              example: not_found
            message:
              type: string
              description: Human-readable explanation.
            request_id:
              type: string
              description: Correlation ID; match against server logs.
            details:
              type: array
              description: Field-level problems (e.g. validation failures).
              items:
                $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        field:
          type: string
          description: Dotted path to the offending field, when applicable.
        message:
          type: string
          description: What is wrong with this field.
  responses:
    Error:
      description: Error response.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RuntimeUnauthorized:
      description: >-
        Uniform machine-channel authentication refusal. The response must not reveal whether the bearer,
        organization, device, agent kind, or revocation state exists.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"